Two functions, and two increments whose order is a delay slot
func_8001EE44 parked, func_80021480 matched. 656 functions.

func_80021480 took three edits and two of them are new:

- `for (i = 0; i < 10; p++, i++)` matches; `i++, p++` is 39 differences.
  With the counter first gcc emits its increment ahead of the load that
  starts the body and the load's delay slot needs a nop. The relative
  order of the two decides which one the scheduler has available there.
  Nothing else in the function moves.
- func_80040410's second parameter was `u8`, which forces an andi at
  every call site; retail has none. The narrowness belonged to the byte
  store inside the callee, which truncates on its own. Widening it to
  s32 left the callee and its other decompiled caller byte-identical --
  checked both with try_func before trusting it -- and removed the andi.
- `tbl[i + K]` emits `addu index,base`; `(tbl + i)[K]` emits
  `addu base,index`. That was the last difference.

func_8001EE44 parks at two registers: retail leaves the extracted field
in $v0 and writes the argument register at the mask, we coalesce the
shift into $a0 and mask in place. Every way of splitting the unmasked
value into its own local costs a register and takes it to 52 -- including
the conditional-expression form that fixed func_800595C8's identically
shaped symptom yesterday, which is worth knowing is not general.

+++ src/func_80040410.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-void func_80040410(u8 *arg0, u8 arg1) {
+void func_80040410(u8 *arg0, s32 arg1) {
     arg0[0x69] = arg1;
     *(u16 *)(arg0 + 8) &= 0xFFEF;
 }
